requirements.txt
pyTelegramBotAPI==4.14.0
google-adk==0.1.0
cryptography>=42.0.0
pyautogui==0.9.54
opencv-python>=4.9.0.80
numpy>=1.26.0
requests>=2.31.0
mss
setuptools>=65.0.0
deprecated>=1.2.0